# src/JsonSchema.Api.Tests/openapi/JsonSchema.Api.Tests.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JsonSchema.Api.Tests | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/test/simple": {
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test/strict": {
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrictModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StrictModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StrictModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test/multiword": {
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiWordModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiWordModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MultiWordModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test/unvalidated": {
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnvalidatedModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnvalidatedModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnvalidatedModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MultiWordModel": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "SimpleModel": {
        "required": [
          "name",
          "age"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "StrictModel": {
        "required": [
          "title",
          "active"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "UnvalidatedModel": {
        "required": [
          "description"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Test"
    }
  ]
}